Updates 'pdf2png' method. No output dir created by default.
'pdf2png' method checks if the PDF file has whitespace in the filename and
renames the file without whitespace. Then it creates the PNG files and
exits. Previous behavior was to not exit or check for whitespaces.

Renamed the 'setup' method to 'output_dir' and removed it from the
'initialize' method. The 'clean' and 'pdf2png' methods didn't need the 'setup'
function run before execution, or at all. 'output_dir' method now called
in the 'run' method, after the check for running 'clean' and 'pdf2png' methods.